Fixes a bug where "this" project wasn't being skipped.
The purpose of the skip is because GetReferencedProjects returns a collection that includes the current project.  We don't want to try to import .ts files from the imported project so we need to skip it.

Previously, there was a bug where we would skip the first found .ts file, not the first project.