Added in a new property "UseSourceDirectoryStructure". If this is set to true then the code will rather than copying directly to LibraryDirectoryFullPath will add in the folder structure where the typescript was found. So if the typescript is in a "Script/Module1" folder inside the typescript project then it will be copied to a Script/Module1 folder inside the target directory. This is useful for maintaining structure, for if you have multiple typescript files with the same name and if you want to preserve a folder structure for other reasons (such as to make sure references between typescript files will stay working as you expect).
To do this the LINQ to get the typescript was split up. We now enumerate through each project first so we can get the source Project's path so that we can then strip that off of the beginning of sourcedirectory path to get the root relative directory path. The main difference once inside the inner enumeration is that rather than using LibraryDirectoryFullPath property directly a method determines what the destination directory should be based on the value of the new property.