Removed call to launch debugger that should 100% not have been there.
Removed use of Property to determine the destination directory - it was determined that maintaining directory structure should be the standard.

Changed some explicit declarations to implicit to match style guidelines.

Reduced excessive indent of chained method calls.

Added in some "NotNull" method calls and some contracts to try to stop the build complaining. Not overly familiar with their use so not 100% sure this will satisfy all the contracts but I think it will?